tech-topic-model/requirements.txt
spacy-streamlit>=1.0.0a0
streamlit
